Complete autonomous functions.
Add optional parameters to DriveDistance to allow for SLOW or NORMAL
speeds, and LOW or HIGH gear. Default is NORMAL speed and LOW gear. We
automatically drop out of high gear when there's two feet left to drive
in order to reduce overshoot. (If the drive distance is less than two
feet, we never go into high gear.) So we alays end our DriveDistance in
low gear.

Added gyro feedback correction to DriveDistance to keep our robot
driving straight.

Allow for two different speeds when crossing defenses. The Portcullis,
Cheval de Frise, and Low Bar are set to SLOW.

In AutonomousDriveSequence, we use HIGH gear (but SLOW speed -- 0.5 for
now) when driving from our crossing point to the target line-up point.

Implement AutonomousDriveAndShoot, which is an entire autonomous. This
does anAutonomousDriveSequence followed by an AutonomousShootSequence.

Add test buttons for AutonomousDriveAndShoot (12) and
AutonomousShootSequence (14) to the Farm controller.